fix: segregate OpenFin snapshots for credit vs FX...
Local storage is shared by both apps, thus without this change, they see each
other's snapshots, which lead to undesirable layout state changes.
This commit appends the app name to the local storage keys to  ensure that the
OpenFin snapshots are segregated across Credit and FX.

Interestingly, on Mac, local storage is segregated for the 2 apps, so
this fix would seem to be unnecessary. On Windows, however, local storage
is shared across apps that are served from the same origin, even if they
have different UUIDs. Creating separate security realms for each app would
create segregated local storage, but has the side effect of breaking
interop with ReactiveAnalytics, so we did not pursue that approach.